requirements.txt
# =========================
# Core framework
# =========================
fastapi==0.115.4
uvicorn==0.30.1
starlette==0.41.2
python-dotenv==1.0.1
protobuf>=4.25.3

# =========================
# Networking
# =========================
httpx>=0.28.1,<1.0.0
websockets>=13.0,<15.0
anyio>=4.0.0
sniffio>=1.3.0
requests>=2.31.0
psycopg2-binary==2.9.9

# =========================
# File / Document processing
# =========================
aiofiles==25.1.0
pillow==10.4.0
PyMuPDF==1.23.26
pdfplumber==0.11.0
pdfminer.six==20231228
pytesseract==0.3.10
python-docx==1.1.2
reportlab==4.1.0
PyPDF2==3.0.1
pypdfium2==4.24.0
pdf2image==1.17.0
python-multipart==0.0.6

# =========================
# Utilities
# =========================
tqdm==4.66.4
PyYAML==6.0.1

# =========================
# AI/ML
# =========================
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.38.0
sentencepiece>=0.2.0
google-genai
google-generativeai
sounddevice
google-cloud-speech
grpcio